+++ lib/sidekiq/transaction_guard/railtie.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Sidekiq::TransactionGuard
+  class Railtie < ::Rails::Railtie
+    initializer "sidekiq.transaction_guard" do
+      Sidekiq::TransactionGuard.mode = (Rails.env.local? ? :error : :warn)
+
+      Sidekiq.configure_client do |config|
+        config.client_middleware do |chain|
+          chain.add Sidekiq::TransactionGuard::Middleware
+        end
+      end
+    end
+  end
+end