+++ spec/spec_helper.rb
@@ -2,16 +2,18 @@
 
 require "bundler/setup"
 require "sidekiq/transaction_guard"
+require "sidekiq/transaction_guard/rspec"
 
 require "active_record"
 
-RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+Sidekiq.logger.level = :error
 
-  config.expect_with :rspec do |c|
-    c.syntax = :expect
-  end
+RSpec.configure do |config|
+  config.warnings = true
+  config.disable_monkey_patching!
+  config.default_formatter = "doc" if config.files_to_run.one?
+  config.order = :random
+  Kernel.srand config.seed
 end
 
 ActiveRecord::Base.establish_connection("adapter" => "sqlite3", "database" => ":memory:")